fix order of when populate is called, fix for premundane dates
